core: add AbstractYieldOperation and use in riscv-scf
